requirements/dev.txt
#Werkzeug required by django-extensions
Werkzeug
django-extensions
django-debug-toolbar
django-tastypie
mimeparse